refactor(map-styles): mv CDN URL fn to map-styles
