+++ packages/components/map/src/composables/openLayersMap.js
@@ -9,11 +9,6 @@ import { useGeographic } from "ol/proj.js";
 import { apply as applyMapboxStyle } from "ol-mapbox-style";
 import LayerGroup from "ol/layer/Group.js";
 
-import {
-  EUROPEANA_MAP_STYLE_NAMES,
-  useEuropeanaMapStyle,
-} from "./europeanaMapStyle.js";
-
 const projection = "EPSG:3857";
 const centreOfEurope = [9.254419, 50.102223];
 
@@ -29,7 +24,7 @@ export const useOpenLayersMap = ({
   const centreRef = toRef(unref(centre) || centreOfEurope);
   const hashRef = toRef(hash);
   const mapRef = toRef(map);
-  const styleRef = toRef(style || "openstreetmap");
+  const styleRef = toRef(style);
   const targetRef = toRef(target);
   // unref first in case it's a computed and we need to set it
   const zoomRef = toRef(unref(zoom) || 4);
@@ -38,27 +33,11 @@ export const useOpenLayersMap = ({
   useGeographic();
 
   const createLayer = () => {
-    let styleNameOrURL;
-    let styleOptions;
-    let styleURL;
-
-    if (Array.isArray(styleRef.value)) {
-      styleNameOrURL = styleRef.value[0];
-      styleOptions = styleRef.value[1];
-    } else {
-      styleNameOrURL = styleRef.value;
-    }
-
-    if (styleNameOrURL === "openstreetmap") {
+    if (!styleRef.value || styleRef.value === "openstreetmap") {
       return new TileLayer({ source: new OSM() });
-    } else if (styleNameOrURL) {
-      if (EUROPEANA_MAP_STYLE_NAMES.includes(styleNameOrURL)) {
-        styleURL = useEuropeanaMapStyle(styleNameOrURL, styleOptions);
-      } else {
-        styleURL = styleNameOrURL;
-      }
+    } else {
       const layerGroup = new LayerGroup();
-      applyMapboxStyle(layerGroup, styleURL);
+      applyMapboxStyle(layerGroup, styleRef.value);
       return layerGroup;
     }
   };
