--- packages/style/map-styles/bin/cli.js
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-
-import { runTask } from "../src/index.js";
-
-runTask(process.argv[2], process.argv[3]);
