+++ packages/style/map-styles/bin/generate.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+import { runTask } from "../src/tasks.js";
+
+runTask("generate", process.argv[2]);
