+++ packages/style/map-styles/package.json
@@ -6,21 +6,23 @@
     "type": "git",
     "url": "https://github.com/europeana/sea"
   },
-  "bin": "./bin/cli.js",
   "license": "EUPL-1.2",
   "type": "module",
   "files": [
-    "dist",
-    "src"
+    "dist"
   ],
   "exports": {
-    "./protomaps": "./src/protomaps/index.js",
-    "./versatiles": "./src/protomaps/versatiles.js"
+    "import": "./dist/index.mjs",
+    "require": "./dist/index.cjs"
   },
   "scripts": {
-    "build": "pnpm build:protomaps && pnpm build:versatiles",
-    "build:protomaps": "pnpm cli protomaps build",
-    "build:versatiles": "pnpm cli versatiles build",
-    "cli": "./bin/cli.js"
+    "build": "pnpm unbuild && pnpm generate",
+    "generate": "pnpm generate:protomaps && pnpm generate:versatiles",
+    "generate:protomaps": "node bin/generate.js protomaps",
+    "generate:versatiles": "node bin/generate.js versatiles",
+    "test": "vitest"
+  },
+  "devDependencies": {
+    "unbuild": "^3.6.1"
   }
 }
