+++ packages/style/map-styles/src/styles/index.js
@@ -1,0 +1,7 @@
+import * as protomaps from "./protomaps/index.js";
+import * as versatiles from "./versatiles/index.js";
+
+export const styles = {
+  protomaps,
+  versatiles,
+};
