+++ packages/style/map-styles/src/styles/protomaps/style.json
