+++ packages/style/map-styles/src/styles/versatiles/style.json
